+++ tests/integration/thrift/client.cc
@@ -27,7 +27,7 @@ namespace Test {
 /* INPUTS */
 
 int64_t now_ns() {
-  return (int64_t)std::chrono::duration_cast<std::chrono::nanoseconds>(
+  return std::chrono::duration_cast<std::chrono::nanoseconds>(
     std::chrono::system_clock::now().time_since_epoch()).count();
 }
